Refactor: Change error code
