+++ internal/server/server.go
@@ -163,7 +163,7 @@ func (s *Server) initRoutes() {
 	userRoutes.Post("/login", s.userHandler.Login)
 	userRoutes.Patch("/", s.authMiddleware.Auth, s.userHandler.UpdateUserInformation)
 	userRoutes.Get("/verify/", s.userHandler.VerifyEmail)
-	userRoutes.Get("/resetpassword", s.userHandler.ResetPasswordCreate)
+	userRoutes.Get("/resetpassword", s.authMiddleware.Auth, s.userHandler.ResetPasswordCreate)
 	userRoutes.Patch("/newpassword", s.userHandler.ResetPasswordResponse)
 	userRoutes.Get("/me", s.userHandler.GetUserInfo)
 }
